sorta sketch bezier curves

+++ src/main/java/frc/positiontracking/pathfinding/Target.java
@@ -32,7 +32,7 @@ public class Target {
         return targets;
     }
 
-    public static double size() {
+    public static int size() {
         return targets.size();
     }
